Moved tests to tests/variational
